+++ src/components/email-renderer.tsx
@@ -1,14 +1,14 @@
-"use server";
+"use server"
 
-import { render } from "@react-email/components";
+import { render } from "@react-email/components"
 
-import { EmailTemplates } from "@/types";
+import { EmailTemplates } from "@/types"
 
-import PasswordWasResetEmail from "../../emails/password-was-reset";
-import ResetPasswordEmail from "../../emails/reset-password";
-import SupportTicketEmail from "../../emails/support-ticket";
-import TestEmail from "../../emails/test-email";
-import VerifyEmailEmail from "../../emails/verify-email";
+import PasswordWasResetEmail from "../../emails/password-was-reset"
+import ResetPasswordEmail from "../../emails/reset-password"
+import SupportTicketEmail from "../../emails/support-ticket"
+import TestEmail from "../../emails/test-email"
+import VerifyEmailEmail from "../../emails/verify-email"
 
 const emailComponents = {
   TestEmail: TestEmail,
@@ -16,18 +16,18 @@ const emailComponents = {
   ResetPassword: ResetPasswordEmail,
   PasswordWasReset: PasswordWasResetEmail,
   SupportTicket: SupportTicketEmail,
-};
+}
 
 type EmailRendererProps = {
-  template?: EmailTemplates;
-  data: {};
-};
+  template?: EmailTemplates
+  data: {}
+}
 
 export const emailRenderer = async ({
   template,
   data = {},
 }: EmailRendererProps) => {
   const EmailComponents =
-    emailComponents[template as keyof typeof emailComponents];
-  return render(<EmailComponents data={data} />);
-};
+    emailComponents[template as keyof typeof emailComponents]
+  return render(<EmailComponents data={data} />)
+}